requirements.txt
fastapi
uvicorn[standard]
numpy
opencv-python-headless
sort-tracker
python-multipart
ultralytics